# src/webapi/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Description of specifications",
        "title": "Rack of labs API",
        "contact": {
            "name": "API supporter",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "license(Mandatory)",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "0.1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1/",
    "paths": {
        "/ethernet-switch/": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ethernet-switch"
                ],
                "summary": "Get paginated list of switches",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Order by field",
                        "name": "orderBy",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "'asc' or 'desc' for ascending or descending order",
                        "name": "orderDirection",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Searchable value in entity",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Number of entities per page",
                        "name": "pageSize",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dtos.ResponseDataDto"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "allOf": [
                                                {
                                                    "$ref": "#/definitions/dtos.PaginatedListDto"
                                                },
                                                {
                                                    "type": "object",
                                                    "properties": {
                                                        "items": {
                                                            "type": "array",
                                                            "items": {
                                                                "$ref": "#/definitions/dtos.EthernetSwitchDto"
                                                            }
                                                        }
                                                    }
                                                }
                                            ]
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ethernet-switch"
                ],
                "summary": "Create new ethernet switch",
                "parameters": [
                    {
                        "description": "Ethernet switch fields",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dtos.EthernetSwitchCreateDto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dtos.ResponseDataDto"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
        "/ethernet-switch/{id}": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ethernet-switch"
                ],
                "summary": "Get ethernet switch by id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Ethernet switch ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dtos.ResponseDataDto"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dtos.EthernetSwitchDto"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ethernet-switch"
                ],
                "summary": "Updates ethernet switch by id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Ethernet switch ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Ethernet switch fields",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dtos.EthernetSwitchUpdateDto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dtos.ResponseDto"
                        }
                    }
                }
            },
            "delete": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ethernet-switch"
                ],
                "summary": "Delete ethernet switch by id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Ethernet switch ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dtos.ResponseDto"
                        }
                    }
                }
            }
        },
        "/ethernet-switch/{id}/port/": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ethernet-switch"
                ],
                "summary": "Gets paginated list of ethernet switch ports",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Ethernet switch ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Order by field",
                        "name": "orderBy",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "'asc' or 'desc' for ascending or descending order",
                        "name": "orderDirection",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Searchable value in entity",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Number of entities per page",
                        "name": "pageSize",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "desc",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dtos.ResponseDataDto"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "allOf": [
                                                {
                                                    "$ref": "#/definitions/dtos.PaginatedListDto"
                                                },
                                                {
                                                    "type": "object",
                                                    "properties": {
                                                        "items": {
                                                            "type": "array",
                                                            "items": {
                                                                "$ref": "#/definitions/dtos.EthernetSwitchPortDto"
                                                            }
                                                        }
                                                    }
                                                }
                                            ]
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ethernet-switch"
                ],
                "summary": "Creates new ethernet switch port",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Ethernet switch ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Ethernet switch port fields",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dtos.EthernetSwitchPortCreateDto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dtos.ResponseDataDto"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/ethernet-switch/{id}/port/{portID}": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ethernet-switch"
                ],
                "summary": "Gets ethernet switch port by id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Ethernet switch ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Ethernet switch port ID",
                        "name": "portID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dtos.ResponseDataDto"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dtos.EthernetSwitchPortDto"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ethernet-switch"
                ],
                "summary": "Updates ethernet switch port by id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Ethernet switch ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Ethernet switch port ID",
                        "name": "portID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Ethernet switch port fields",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dtos.EthernetSwitchPortUpdateDto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dtos.ResponseDto"
                        }
                    }
                }
            },
            "delete": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ethernet-switch"
                ],
                "summary": "Delete ethernet switch port by id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Ethernet switch ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Ethernet switch port ID",
                        "name": "portID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dtos.ResponseDto"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dtos.DtoType": {
            "type": "object"
        },
        "dtos.EthernetSwitchCreateDto": {
            "type": "object",
            "properties": {
                "address": {
                    "description": "Address - switch ip address",
                    "type": "string"
                },
                "name": {
                    "description": "Name - switch name",
                    "type": "string"
                },
                "password": {
                    "description": "Password - ethernet switch management password",
                    "type": "string"
                },
                "serial": {
                    "description": "Serial - switch serial number",
                    "type": "string"
                },
                "switchModel": {
                    "description": "SwitchModel - switch model",
                    "type": "string"
                },
                "username": {
                    "description": "Username - switch admin username",
                    "type": "string"
                }
            }
        },
        "dtos.EthernetSwitchDto": {
            "type": "object",
            "properties": {
                "address": {
                    "description": "Address - switch ip address",
                    "type": "string"
                },
                "createdAt": {
                    "description": "CreatedAt - entity create time",
                    "type": "string"
                },
                "id": {
                    "description": "ID - unique identifier",
                    "type": "string"
                },
                "name": {
                    "description": "Name - switch name",
                    "type": "string"
                },
                "serial": {
                    "description": "Serial - switch serial number",
                    "type": "string"
                },
                "switchModel": {
                    "description": "SwitchModel - switch model",
                    "type": "string"
                },
                "updatedAt": {
                    "description": "UpdatedAt - entity update time",
                    "type": "string"
                },
                "username": {
                    "description": "Username - switch admin username",
                    "type": "string"
                }
            }
        },
...
        "dtos.EthernetSwitchPortCreateDto": {
            "type": "object",
            "properties": {
                "name": {
                    "description": "Name for this port",
                    "type": "string"
                },
                "poetype": {
                    "description": "POEType type of PoE for this port\ncan be: \"poe\", \"poe+\", \"passive24\", \"none\"",
                    "type": "string"
                }
            }
        },
        "dtos.EthernetSwitchPortDto": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "description": "CreatedAt - entity create time",
                    "type": "string"
                },
                "id": {
                    "description": "ID - unique identifier",
                    "type": "string"
                },
                "name": {
                    "description": "Name for this port",
                    "type": "string"
                },
                "poetype": {
                    "description": "POEType type of PoE for this port\ncan be: \"poe\", \"poe+\", \"passive24\", \"none\"",
                    "type": "string"
                },
                "updatedAt": {
                    "description": "UpdatedAt - entity update time",
                    "type": "string"
                }
            }
        },
        "dtos.EthernetSwitchPortUpdateDto": {
            "type": "object",
            "properties": {
                "name": {
                    "description": "Name for this port",
                    "type": "string"
                },
                "poetype": {
                    "description": "POEType type of PoE for this port\ncan be: \"poe\", \"poe+\", \"passive24\", \"none\"",
                    "type": "string"
                }
            }
        },
        "dtos.EthernetSwitchUpdateDto": {
            "type": "object",
            "properties": {
                "address": {
                    "description": "Address - switch ip address",
                    "type": "string"
                },
                "name": {
                    "description": "Name - switch name",
                    "type": "string"
                },
                "password": {
                    "description": "Password - ethernet switch management password",
                    "type": "string"
                },
                "serial": {
                    "description": "Serial - switch serial number",
                    "type": "string"
                },
                "switchModel": {
                    "description": "SwitchModel - switch model",
                    "type": "string"
                },
                "username": {
                    "description": "Username - switch admin username",
                    "type": "string"
                }
            }
        },
...
        "dtos.PaginatedListDto": {
            "type": "object",
            "properties": {
                "items": {
                    "description": "Items - array of DTO items",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dtos.DtoType"
                    }
                },
                "page": {
                    "description": "Page - page number",
                    "type": "integer"
                },
                "size": {
                    "description": "Size - page size",
                    "type": "integer"
                },
                "total": {
                    "description": "Total - total number of items",
                    "type": "integer"
                }
            }
        },
        "dtos.ResponseDataDto": {
            "type": "object",
            "properties": {
                "data": {
                    "description": "Data - any data to send"
                },
                "status": {
                    "description": "Status - response status structure",
                    "$ref": "#/definitions/dtos.ResponseStatusDto"
                }
            }
        },
        "dtos.ResponseDto": {
            "type": "object",
            "properties": {
                "status": {
                    "description": "Status - response status structure",
                    "$ref": "#/definitions/dtos.ResponseStatusDto"
                }
            }
        },
        "dtos.ResponseStatusDto": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "Code - 0 if OK, otherwise 1",
                    "type": "integer"
                },
                "message": {
                    "description": "Message - response message",
                    "type": "string"
                }
            }
        }
    }
}